requirements.txt
beautifulsoup4
celery[redis]
csscompressor
dj-database-url
Django
djangorestframework
django-autoslug
django-celery-beat
django-celery-results
django-filter
django_compressor
gunicorn
html5lib
markdown
psycopg[binary,pool]
requests
whitenoise[brotli]
